Add separate implementation for Kotlin and Java
When the Kotlin version (or previous version) of this library is used in
a Java project some transitive Kotlin dependencies are also added to the
project. For a library that small this should not happen.
On the other hand a Kotlin application may want to profit from
null-safety Kotlin offers.